+++ test/selectors/optimizer-test.js
@@ -9,6 +9,7 @@ function optimizerContext(group, specs, options) {
 
   var context = {};
   options = options || {};
+  options.shorthandCompacting = true;
   options.compatibility = new Compatibility(options.compatibility).toOptions();
 
   function optimized(target) {